+++ src/proofOfWorkNodeJs.ts
@@ -1,9 +1,8 @@
-import { ArrayHelper } from "@iota-pico/core/dist/helpers/arrayHelper";
 import { NumberHelper } from "@iota-pico/core/dist/helpers/numberHelper";
 import { ObjectHelper } from "@iota-pico/core/dist/helpers/objectHelper";
+import { ITimeService } from "@iota-pico/core/dist/interfaces/ITimeService";
 import { CryptoError } from "@iota-pico/crypto/dist/error/cryptoError";
-import { IProofOfWork } from "@iota-pico/crypto/dist/interfaces/IProofOfWork";
-import { Hash } from "@iota-pico/data/dist/data/hash";
+import { ProofOfWorkBase } from "@iota-pico/crypto/dist/proofOfWork/proofOfWorkBase";
 import { Trytes } from "@iota-pico/data/dist/data/trytes";
 import * as ffi from "ffi";
 import * as fs from "fs";
@@ -15,7 +14,7 @@ import { INodePlatform } from "./INodePlatform";
 /**
  * ProofOfWork implementation using NodeJS.
  */
-export class ProofOfWorkNodeJs implements IProofOfWork {
+export class ProofOfWorkNodeJs extends ProofOfWorkBase {
     /* @internal */
     private readonly _nodePlatform: INodePlatform;
 
@@ -29,8 +28,10 @@ export class ProofOfWorkNodeJs implements IProofOfWork {
     /**
      * Create a new instance of ProofOfWork.
      * @param nodePlatform Provides platform specific functions, optional mostly used for testing.
+     * @param timeService Service to get the time for attachments.
      */
-    constructor(nodePlatform?: INodePlatform) {
+    constructor(nodePlatform?: INodePlatform, timeService?: ITimeService) {
+        super(timeService);
         if (ObjectHelper.isEmpty(nodePlatform)) {
             this._nodePlatform = {
                 pathResolve: path.resolve,
@@ -49,6 +50,8 @@ export class ProofOfWorkNodeJs implements IProofOfWork {
      * Will throw an exception if the implementation is not supported.
      */
     public async initialize(): Promise<void> {
+        await super.initialize();
+
         const fullPath = this._nodePlatform.pathJoin(path.join(__dirname, "../binaries/"));
 
         const platform = this._nodePlatform.platform();
@@ -71,37 +74,27 @@ export class ProofOfWorkNodeJs implements IProofOfWork {
     }
 
     /**
-     * Performs single conversion per pow call.
-     * @returns True if pow only does one conversion.
-     */
-    public performsSingle(): boolean {
-        return true;
-    }
-
-    /**
-     * Perform a proof of work on the data.
-     * @param trunkTransaction The trunkTransaction to use for the pow.
-     * @param branchTransaction The branchTransaction to use for the pow.
+     * Perform a proof of work on a single item.
      * @param trytes The trytes to perform the pow on.
      * @param minWeightMagnitude The minimum weight magnitude.
      * @returns The trytes produced by the proof of work.
      */
-    public async pow(trunkTransaction: Hash, branchTransaction: Hash, trytes: Trytes[], minWeightMagnitude: number): Promise<Trytes[]> {
-        return new Promise<Trytes[]>((resolve, reject) => {
+    public async singlePow(trytes: Trytes, minWeightMagnitude: number): Promise<Trytes> {
+        return new Promise<Trytes>((resolve, reject) => {
             if (ObjectHelper.isEmpty(this._library)) {
                 throw new CryptoError("Library not loaded, have you called initialize");
             }
-            if (!ArrayHelper.isTyped(trytes, Trytes)) {
-                throw new CryptoError("The trytes must be an array of type Trytes");
+            if (!ObjectHelper.isType(trytes, Trytes)) {
+                throw new CryptoError("The trytes must be an object of type Trytes");
             }
             if (!NumberHelper.isInteger(minWeightMagnitude) || minWeightMagnitude <= 0) {
                 throw new CryptoError("The minWeightMagnitude must be > 0");
             }
-            this._library.ccurl_pow.async(trytes[0].toString(), minWeightMagnitude, (error, returnedTrytes) => {
+            this._library.ccurl_pow.async(trytes.toString(), minWeightMagnitude, (error, returnedTrytes) => {
                 if (error) {
                     reject(error);
                 } else {
-                    resolve([ Trytes.fromString(returnedTrytes) ]);
+                    resolve(Trytes.fromString(returnedTrytes));
                 }
             });
         });