+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.9.2
+
+* Can now perform multi-pass pow as well as single pass
+* Added CommonJS version of library in pkg folder
+
 ## v0.9.1
 
 * Fix Release